requirements.txt
# 项目依赖
pillow>=10.0.0
easyocr>=1.7.0
numpy>=1.24.0
opencv-python>=4.8.0

